backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.14.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
instructor==1.14.1
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
nodeenv==1.10.0
openai==2.15.0
platformdirs==4.5.1
praw==7.8.1
prawcore==2.4.0
pre_commit==4.5.1
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
ty==0.0.11
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.1
websocket-client==1.9.0
yarl==1.22.0
